linebot_message/requirements.txt
Flask==3.0.3
line-bot-sdk==3.11.0
python-dotenv==1.0.1
openai==1.37.1
httpx==0.27.2